Report changes in memory usage that would result from merging a PR
On creation or commit to a pull request, a report of the resulting change in memory usage of the examples will be
commented to the PR thread.

NOTE: this version of the workflow is designed for use in private repositories, as explained here:

https://github.com/arduino/report-size-deltas#run-from-the-same-workflow-as-the-arduinocompile-sketches-action

Once the repository is made public, it must be replaced by these workflows:

- https://github.com/per1234/.github/blob/main/workflow-templates/compile-examples.md
- https://github.com/per1234/.github/blob/main/workflow-templates/report-size-deltas.md

The reason for this is the private repository version will fail on pull requests submitted from forks due to the
automatic token permissions downgrade done by GitHub actions under those conditions for security purposes.