Switch to variant of "Compile Examples" workflow for public repos (#56)
Initial development of the project was done in a private repository. In that environment, the most efficient
configuration of the workflow is done by running the `arduino/report-size-deltas` action from a job in the same
workflow.

However, that configuration is not suitable for use in a public repository. The reason is that the deltas report comment
requires a GitHub access token with write permissions and the token does not have such permissions when the
"Compile Examples" workflow is triggered by events in a pull request from a fork. For this reason, it is necessary to
run the `arduino/report-size-deltas` action from a dedicated scheduled workflow, which will always have write
permissions.